# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Partial 1 - Rugby World Cup 2023'","description":"'Partial 1 - Rugby World Cup 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Partial 1 - Rugby World Cup 2023'"}],"paths":{"/rwc/2023/results/{group}":{"get":{"tags":["pool-controller"],"operationId":"getResultPool","parameters":[{"name":"group","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PoolResultsDto"}}}}}}},"/rwc/2023/pools":{"get":{"tags":["pool-controller"],"operationId":"getAllPools","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PoolDto"}}}}}}}},"/rwc/2023/pool/{group}":{"get":{"tags":["pool-controller"],"operationId":"getPoolById","parameters":[{"name":"group","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PoolDto"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"PoolResultsDto":{"type":"object","properties":{"poolId":{"type":"string"},"teams":{"type":"array","items":{"$ref":"#/components/schemas/TeamStanding"}}}},"TeamStanding":{"type":"object","properties":{"teamId":{"type":"integer","format":"int64"},"teamName":{"type":"string"},"country":{"type":"string"},"matchesPlayed":{"type":"integer","format":"int32"},"wins":{"type":"integer","format":"int32"},"draws":{"type":"integer","format":"int32"},"losses":{"type":"integer","format":"int32"},"pointsFor":{"type":"integer","format":"int32"},"pointsAgainst":{"type":"integer","format":"int32"},"pointsDifferential":{"type":"integer","format":"int32"},"triesMade":{"type":"integer","format":"int32"},"bonusPoints":{"type":"integer","format":"int32"},"points":{"type":"integer","format":"int32"},"totalYellowCards":{"type":"integer","format":"int32"},"totalRedCards":{"type":"integer","format":"int32"}}},"PoolDto":{"type":"object","properties":{"pool":{"type":"string"},"teams":{"type":"array","items":{"$ref":"#/components/schemas/Team"}}}},"Team":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"country":{"type":"string"},"worldRanking":{"type":"integer","format":"int32"},"pool":{"type":"string"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}}}}}