# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/PoolDto.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import ar.edu.utn.frc.tup.lciii.models.Team;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class PoolDto {
    private String pool;
    private List<Team> teams;
}