# src/main/java/ar/edu/utn/frc/tup/lciii/models/Team.java
package ar.edu.utn.frc.tup.lciii.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Team {
    private Long id;
    private String name;
    private String country;
    private Integer worldRanking;
    private String pool;
}